+++ plaso/output/null.py
@@ -1,18 +1,18 @@
 # -*- coding: utf-8 -*-
-"""An output module that doesn't output anything."""
+"""Null device output module."""
 
 from plaso.output import interface
 from plaso.output import manager
 
 
 class NullOutputModule(interface.OutputModule):
-  """An output module that doesn't output anything."""
+  """Null device output module."""
 
   NAME = u'null'
-  DESCRIPTION = u'An output module that doesn\'t output anything.'
+  DESCRIPTION = u'Output module that does not output anything.'
 
-  def WriteEventBody(self, event_object):
-    """Writes the body of an event object to the output.
+  def WriteEventBody(self, unused_event_object):
+    """Writes the body of an event object to a null device.
 
     Args:
       event_object: the event object (instance of EventObject).
