+++ tests/multi_processing/engine.py
@@ -35,38 +35,5 @@ class MultiProcessEngineTest(shared_test_lib.BaseTestCase):
   # TODO: add test for _TerminateProcess
 
 
-class TaskBasedMultiProcessEngineTest(shared_test_lib.BaseTestCase):
-  """Tests for the task-based multi-process engine."""
-
-  def testProcessSources(self):
-    """Tests the PreprocessSources and ProcessSources function."""
-    test_engine = engine.TaskBasedMultiProcessEngine(
-        maximum_number_of_tasks=100)
-
-    source_path = os.path.join(self._TEST_DATA_PATH, u'ímynd.dd')
-    os_path_spec = path_spec_factory.Factory.NewPathSpec(
-        dfvfs_definitions.TYPE_INDICATOR_OS, location=source_path)
-    source_path_spec = path_spec_factory.Factory.NewPathSpec(
-        dfvfs_definitions.TYPE_INDICATOR_TSK, location=u'/',
-        parent=os_path_spec)
-
-    test_engine.PreprocessSources([source_path_spec])
-
-    session = sessions.Session()
-
-    with shared_test_lib.TempDirectory() as temp_directory:
-      temp_file = os.path.join(temp_directory, u'storage.plaso')
-      storage_writer = storage_zip_file.ZIPStorageFileWriter(
-          session, temp_file)
-
-      test_engine.ProcessSources(
-          session.identifier, [source_path_spec], storage_writer,
-          parser_filter_expression=u'filestat')
-
-    # TODO: implement a way to obtain the results without relying
-    # on multi-process primitives e.g. by writing to a file.
-    # self.assertEqual(len(storage_writer.events), 15)
-
-
 if __name__ == '__main__':
   unittest.main()
